+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 
 ### Bug Fixes
 
-* include download urls in userscript metadata ([#2](https://github.com/popstas/userscripts/issues/2)) ([92a467f](https://github.com/popstas/userscripts/commit/92a467f431f3d3eba513e97d8e245509430657c6))
+* include download urls in userscript metadata (#2) 92a467f, closes #2
+
+
+### Features
+
+* support single task status in codex-helper 9230b19